requirements.txt
# Core ML packages
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.3.2

# Django and extensions
Django==5.0.6
django-registration==3.4
django-otp==1.5.4
django-two-factor-auth==1.15.5
djangorestframework==3.14.0
django-simple-history==3.4.0
django-cleanup==8.1.0
django-compressor==4.4
django-sass-processor==1.4
django-redis==5.4.0
django-allauth==0.61.1
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-cors-headers==4.3.1
django-storages==1.14.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Image processing and payment
Pillow==10.2.0
stripe==7.11.0

# NLP and chatbot
spacy==3.7.4
symspellpy==6.7.0
rasa==3.6.15

# Testing and development
pytest==8.0.1
pytest-django==4.8.0
pytest-cov==4.1.0
coverage==7.4.1
factory-boy==3.3.0
Faker==23.2.1

# Task queue and caching
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Web server and performance
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9

# Real-time features
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
uvicorn==0.27.1

# Utility packages
python-dotenv==1.0.1
boto3==1.34.34
ipython==8.22.2
requests==2.31.0
python-dateutil==2.8.2
libsass==0.22.0
phonenumbers==9.0.2 